requirements.txt
langchain
langchain_core
langchain_community
langchain_groq
langchain_huggingface
pandas
parsivar
python-dotenv
streamlit


